`months._load_yearly_parquets`: cast `year`/`month_idx` to int64
On GHA (pandas 3.0) the merge against `day_hists` (int64 index) silently
produced NaN for all rows because the yearly parquet's `year`/`month_idx`
were int32. Pandas 3.0 no longer auto-coerces int32↔int64 in merges.
Locally (pandas 2.3) the merge worked.

Cast to int64 explicitly in `_load_yearly_parquets` so the merge keys
align in both versions. Verified `data/all.pqt` still produces MD5
`6b50b2772e…` (byte-identical).

+++ path_data/months.py
@@ -52,8 +52,10 @@ def _load_yearly_parquets() -> pd.DataFrame:
     df['station'] = df.station.apply(lambda s: RENAME_STATIONS.get(s, s))
     df = df[~df.station.str.contains('TOTAL')].reset_index(drop=True)
     df['dt'] = df['month']
-    df['year'] = df.dt.dt.year
-    df['month_idx'] = df.dt.dt.month
+    # Cast to int64: the day_hists index is int64 and pandas 3.0 no longer
+    # auto-coerces int32↔int64 in merge/join, silently producing NaN.
+    df['year'] = df.dt.dt.year.astype('int64')
+    df['month_idx'] = df.dt.dt.month.astype('int64')
     return df
